added test

+++ src/gen2/util/Functions.java
@@ -6,14 +6,13 @@ public class Functions {
         return 1 / (1 + Math.exp(-x));
     }
 
-
     public static Object getRandom(Object[] col) {
         return col[(int) (Math.random() * col.length)];
     }
 
-    public static float[] convolveCircularly (float[] array, float[] filter) {
+    public static double[] convolveCircularly (double[] array, double[] filter) {
         int sz = array.length, fsz = filter.length;
-        float[] out = new float[sz];
+        double[] out = new double[sz];
         for (int i = 0; i < sz; i++) {
             for (int j = 0; j < fsz; j++) {
                 out[i] += filter[j] * array[Math.floorMod(j-fsz/2+i, sz)];
